pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Lucas <lucasmiguelmac@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
starlette = "0.16"
six = "^1.16.0"
strawberry-graphql = {extras = ["debug-server"], version = "^0.95.0"}
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.11.0"
gunicorn = "^20.1.0"
httptools = "^0.3.0"
uvloop = "^0.16.0"
strawberry-django-plus = "^1.1.2"
typing_extensions = "^4.1.1"

[tool.poetry.dev-dependencies]
django-simple-history = "^3.0.0"
django-extensions = "^3.1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
